SmartMenuBackend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
google-cloud-vision==3.4.0
gunicorn==20.1.0
strsimpy==0.2.1
werkzeug==2.2.3 